Refactor app to backend lib and app
