+++ ft_lib/src/log.cpp
