+++ ft_lib/src/log.hpp
