+++ ft_lib/src/macros.hpp
