+++ ft_lib/src/print.cpp
@@ -1,6 +1,6 @@
-#include "print.hpp"
+#include <ft/print.hpp>
 
-#include "match.hpp"
+#include <ft/match.hpp>
 
 #include <iostream>
